--- js/functions.js
@@ -1,19 +1,0 @@
-let lineLength = (string, length) => {
-  if (string.length <= length) {
-    return true
-  }
-  return false
-};
-
-let palindrome = (string) => {
-  let convertedString = string.replaceAll(' ', '').toLowerCase();
-  for (let i = 0; i < convertedString.length / 2; i++) {
-    if (convertedString[i] != convertedString[convertedString.length - 1 - i]) {
-      return false
-    }
-  }
-  return true
-};
-
-lineLength('Строка', 10);
-palindrome('Лёша на полке клопа нашёл ');
